+++ src/md5.h
@@ -8,13 +8,12 @@
 #ifndef md5_h
 #define md5_h
 
-#include <lua.h>
-
 
 #define HASHSIZE       16
 
+
 void md5 (const char *message, long len, char *output);
-int luaopen_md5_core (lua_State *L);
+LUALIB_API int luaopen_md5 (lua_State *L);
 
 
 #endif
